Adjust for wp-config changes in the WordPress image.
None of the overrides worked anymore in the latest docker image because overriding requires a special wp-config.php, which we now use.
See https://github.com/docker-library/wordpress/issues/576

The previous values that need to be preserved are now extracted to a environment file for Docker.

A future commit might also include username, password and the database values, which for now stay hard-coded.